chore: Add conditional rendering for fundraising cards
This commit adds conditional rendering logic to the `+page.svelte` file to display a message when there are no fundraising cards available. If the `data.props.galangs` array is empty, the message "Belum Ada Penggalangan Dana" is shown. Otherwise, the cards are rendered using a loop through the `data` array. Each card includes an image, title, description, and a button to view more details.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (chore for non-functional changes, feat for new features) and provides a clear and concise description of the code changes.